Add files via upload
A few things to note: I found 499996 to be the delay for the LED to be on which yields around a 20% duty cycle with 8 Hz. By the time 499996 gets to 0 by subtracting 1 every time, around 25 milliseconds have passed, which is 20% duty cycle with 8Hz. Register 2 is the number we multiply 499996 by to get to 40%, 60%, 80%, 100%, and 0% duty cycle. Also, I used Branch Greater Than instead of Branch Not Equal to because when the duty cycle was 0, I subtracted 1 to get to -1, and after I'm in the negatives, I never get back to 0, so I changed it from BNE. 

As of right now, there is an incredibly small spike when the duty cycle is 0 or 100% simply because I turn the LED on or off, THEN check how long I should wait before toggling it again. So, to fix this, we might want to add a check before we toggle the LED to see if the duty cycle is 0 or 100. I did not yet implement the switch because I realized that the way I was doing the lab before was flawed and I needed to change it (I was making R2 the actual delay for the LED to be on (499996 for a 20% duty cycle) instead of the multiple of it. This meant that I could not have 100% duty cycle and a 0% duty cycle using the modulo operator (5*499996 = 2999980, which should be 100% duty cycle, but the next one after that should be 0%. If I did a modulo with 2999980, I could never get to 100% duty cycle since 2999980%2999980 = 0. This presented a problem, so I changed R2 to the MULTIPLE of 499996 so I could do R2 = (R2 + 1)%6, and when R2 = 5, the LED is always on, and when R2 = 6%6 = 0, the LED is off).